Add support for bypassing harvesting of CKAN datasets on DGU
This is useful for remote instance to define if for instance they have
datasets that could potentially get duplicated on DGU.

A new check has been added at the fetch stage that checks if a
dgu_harvest_me extra is present and the value is false. If that's the
case the import stage is skipped completely.

Even though the current CKAN harvester uses the old v2 REST API
(where extras are a dict), support for v3 datasets (extras as a list of
dicts) has been added to future-proof it.